+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: manujime <manujime@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 11:41:06 by manujime          #+#    #+#             */
-/*   Updated: 2022/11/23 16:40:28 by manujime         ###   ########.fr       */
+/*   Updated: 2022/11/24 12:42:21 by manujime         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ int		ft_printchar(int c);
 int		ft_printstr(char *str);
 int		ft_printnbr(int nbr);
 int		ft_printptr(void *ptr);
-int		ft_printhex(unsigned int nbr, char f);
+int		ft_printhex(unsigned long nbr, char f);
 
 #endif
